\name{MetaMultiSKAT-package}
\alias{MetaMultiSKAT-package}
\alias{MetaMultiSKAT}
\docType{package}
\title{
\packageTitle{MetaMultiSKAT}
}
\description{
\packageDescription{MetaMultiSKAT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MetaMultiSKAT}
\packageIndices{MetaMultiSKAT}
}
\author{
\packageAuthor{MetaMultiSKAT}{Diptavo Dutta}

Maintainer: \packageMaintainer{MetaMultiSKAT}
}
\references{
~~ Literature or other references for background information ~~
}
