\name{minP.Meta}
\alias{minP.Meta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
minimum p-value based omnibus test
}
\description{
}
\usage{
minP.Meta(list.MM.objects)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{list.MM.objects}{
	list of Meta.MultiSKAT.wResample objects to be aggregated.
}
}